+++ examples/exposed/src/main/kotlin/UsersSchema.kt
@@ -6,6 +6,7 @@ import org.jetbrains.exposed.dao.id.EntityID
 import org.jetbrains.exposed.dao.id.IntIdTable
 import org.jetbrains.exposed.sql.Column
 import org.jetbrains.exposed.sql.javatime.datetime
+import xyz.daimones.ktor.panel.database.FileUpload
 import java.time.LocalDateTime
 
 enum class Role { SUPER_ADMIN, EDITOR, VIEWER }
@@ -16,6 +17,11 @@ object Users : IntIdTable() {
     val lastName: Column<String> = varchar("last_name", length = 255)
     val password: Column<String> = varchar("password", length = 100)
     val role = enumerationByName("role", 15, Role::class).default(Role.SUPER_ADMIN)
+
+    @FileUpload(storage = "local", path = "/uploads")
+    val image: Column<String> = varchar("image", length = 255)
+    val thumbnail = binary("thumbnail")
+    var blobing = blob("blobing")
     val isActive: Column<Boolean> = bool("is_active").default(false)
     val created: Column<LocalDateTime> = datetime("created").default(LocalDateTime.now())
     val modified: Column<LocalDateTime> = datetime("modified").default(LocalDateTime.now())
@@ -30,6 +36,9 @@ class User(id: EntityID<Int>) : IntEntity(id) {
     var password by Users.password
     var isActive by Users.isActive
     var role by Users.role
+    var image by Users.image
+    var thumbnail by Users.thumbnail
+    var blobing by Users.blobing
     var created by Users.created
     var modified by Users.modified
 }