+++ examples/mongo/src/main/kotlin/Routing.kt
@@ -1,15 +1,16 @@
 package com.example
 
-import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.application.*
-import io.ktor.server.plugins.contentnegotiation.*
+import io.ktor.server.http.content.*
 import io.ktor.server.response.*
 import io.ktor.server.routing.*
+import java.io.File
 
 fun Application.configureRouting() {
     routing {
         get("/") {
             call.respondText("Hello World!")
         }
+        staticFiles("/uploads", File("uploads"))
     }
 }