+++ examples/mongo/src/main/kotlin/UsersSchema.kt
@@ -4,6 +4,7 @@ import kotlinx.serialization.Contextual
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import org.bson.types.ObjectId
+import xyz.daimones.ktor.panel.database.FileUpload
 import xyz.daimones.ktor.panel.database.serialization.LocalDateTimeSerializer
 import java.time.LocalDateTime
 
@@ -19,6 +20,8 @@ data class User(
     val password: String,
     val isActive: Boolean = false,
     val role: Role = Role.SUPER_ADMIN,
+    @FileUpload(storage = "local", path = "/uploads")
+    val image: String? = null,
     @Serializable(with = LocalDateTimeSerializer::class)
     val created: LocalDateTime = LocalDateTime.now(),
     @Serializable(with = LocalDateTimeSerializer::class)