+++ lib/src/main/kotlin/xyz/daimones/ktor/panel/database/dao/ExposedDao.kt
@@ -2,20 +2,16 @@ package xyz.daimones.ktor.panel.database.dao
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import org.jetbrains.exposed.dao.Entity
-import org.jetbrains.exposed.dao.IntEntity
-import org.jetbrains.exposed.dao.IntEntityClass
-import org.jetbrains.exposed.sql.Database
-import org.jetbrains.exposed.sql.SchemaUtils
+import org.jetbrains.exposed.dao.id.EntityID
+import org.jetbrains.exposed.dao.id.IdTable
+import org.jetbrains.exposed.sql.*
+import org.jetbrains.exposed.sql.statements.InsertStatement
+import org.jetbrains.exposed.sql.statements.UpdateStatement
 import org.jetbrains.exposed.sql.transactions.transaction
+import xyz.daimones.ktor.panel.camelToSnake
 import xyz.daimones.ktor.panel.database.DataAccessObjectInterface
-import xyz.daimones.ktor.panel.database.entities.AdminUser
 import xyz.daimones.ktor.panel.database.entities.AdminUsers
-import xyz.daimones.ktor.panel.snakeToCamel
 import kotlin.reflect.KClass
-import kotlin.reflect.KMutableProperty1
-import kotlin.reflect.full.companionObjectInstance
-import kotlin.reflect.full.memberProperties
 
 /**
  * ExposedDao is an implementation of DataAccessObjectInterface using Exposed ORM.
@@ -29,54 +25,71 @@ internal class ExposedDao<T : Any>(
     private val entityKClass: KClass<T>
 ) : DataAccessObjectInterface<T> {
     /** Companion object instance for the entity */
-    private val companion = (entityKClass.companionObjectInstance as? IntEntityClass<IntEntity>)
-        ?: throw IllegalArgumentException("Provided KClass must have a companion object that is an IntEntityClass")
+    private val companion = (entityKClass.objectInstance as? Table)
+        ?: throw IllegalArgumentException("Provided KClass must have an object instance that is a Table.")
 
     /**
-     * Copies properties from the source object to the target IntEntity.
-     * This is a utility function to avoid code duplication in save and update methods.
-     * 
-     * @param source The source object from which to copy properties.
-     * @param target The target IntEntity to which properties will be copied.
+     * Assigns values from a map to an InsertStatement or UpdateStatement.
+     *
+     * @param table The Exposed Table to which the data belongs.
+     * @param data The map containing field names and their corresponding values.
+     * @param insert The InsertStatement to which values will be assigned (if not null).
+     * @param update The UpdateStatement to which values will be assigned (if not null).
      */
-    private fun copyProperties(source: Any, target: Entity<Int>) {
-        val targetProperties = target::class.memberProperties
-            .filterIsInstance<KMutableProperty1<*, *>>()
-            .associateBy { it.name }
+    fun assignMapToInsertOrUpdate(
+        table: Table,
+        data: Map<String, Any>,
+        insert: InsertStatement<Number>? = null,
+        update: UpdateStatement? = null
+    ) {
+        fun populateTable(insert: InsertStatement<Number>? = null, update: UpdateStatement? = null) {
+            for ((key, value) in data) {
+                if (key == "id") continue // skip id if auto-generated
+                if (value == "") continue // skip empty values
+                val column = table.columns.find { it.name == camelToSnake(key) }
+                if (column != null) {
+                    val columnType = column.columnType
+                    val finalValue = when (columnType) {
+                        is EnumerationNameColumnType<*> -> {
+                            val enumClass = columnType.klass.java
+                            java.lang.Enum.valueOf(enumClass, value.toString())
+                        }
 
-        if (source is Map<*, *>) {
-            // If the source is a Map, iterate through its key-value pairs.
-            for ((key, value) in source) {
-                var name = key as String
-                name = snakeToCamel(name)
-                if (name == "id") continue
+                        is EnumerationColumnType<*> -> {
+                            val enumClass = columnType.klass.java
+                            java.lang.Enum.valueOf(enumClass, value.toString())
+                        }
 
-                targetProperties[name]?.let { targetProp ->
-                    @Suppress("UNCHECKED_CAST")
-                    val propClass = (targetProp as KMutableProperty1<Any, Any?>).returnType.classifier as? KClass<*>
-                    val finalValue = if (propClass?.java?.isEnum == true && value is String) {
-                        @Suppress("UNCHECKED_CAST")
-                        java.lang.Enum.valueOf(propClass.java as Class<out Enum<*>>, value)
-                    } else {
-                        value
+                        else -> value
                     }
-                    targetProp.set(target, finalValue)
-                }
-            }
-        } else {
-            // This is the original logic that works for data classes.
-            val sourceProperties = source::class.memberProperties.associateBy { it.name }
-            for ((name, sourceProp) in sourceProperties) {
-                if (name == "id") continue
-
-                targetProperties[name]?.let { targetProp ->
-                    val value = sourceProp.call(source)
                     @Suppress("UNCHECKED_CAST")
-                    (targetProp as KMutableProperty1<Any, Any?>).set(target, value)
+                    if (insert != null) {
+                        (column as? Column<Any>)?.let { insert[it] = finalValue }
+                    } else if (update != null) {
+                        (column as? Column<Any>)?.let { update[it] = finalValue }
+                    }
                 }
             }
         }
+        if (insert != null) {
+            populateTable(insert)
+        } else if (update != null) {
+            populateTable(update = update)
+        }
     }
+
+    /**
+     * Converts a ResultRow to a Map representing the entity.
+     *
+     * @param row The ResultRow to convert.
+     * @return A Map representing the entity.
+     */
+    private fun resultRowToEntity(row: ResultRow): Map<String, Any?> {
+        return row.fieldIndex.keys.associate { col ->
+            (col as Column<*>).name to row[col]
+        }
+    }
+
 
     /**
      * Finds an entity by its primary key.
@@ -84,10 +97,22 @@ internal class ExposedDao<T : Any>(
      * @param id The primary key of the entity to find.
      * @return The found entity of type T, or null if not found.
      */
-    override suspend fun findById(id: Any): T? {
+    override suspend fun findById(id: Any, castToEntityId: Boolean): T? {
+        @Suppress("UNCHECKED_CAST")
+        val idColumn = companion.primaryKey?.columns?.firstOrNull() as? Column<Any>
+            ?: throw IllegalStateException("No primary key defined for table.")
         val result = withContext(Dispatchers.IO) {
             transaction(this@ExposedDao.database) {
-                companion.findById(id.toString().toInt())
+                var entityId: Any = id
+                if (castToEntityId) {
+                    @Suppress("UNCHECKED_CAST")
+                    entityId = EntityID(id as Int, idColumn.table as IdTable<Int>)
+                }
+                companion.selectAll().where { idColumn eq entityId }
+                    .withDistinct()
+                    .map {
+                        resultRowToEntity(it)
+                    }.firstOrNull()
             }
         }
         @Suppress("UNCHECKED_CAST")
@@ -100,9 +125,11 @@ internal class ExposedDao<T : Any>(
      * @return A list of all entities of type T.
      */
     override suspend fun findAll(): List<T?> {
+        companion.primaryKey?.columns?.firstOrNull()
+            ?: throw IllegalStateException("No primary key defined for table.")
         val result = withContext(Dispatchers.IO) {
             transaction(this@ExposedDao.database) {
-                companion.all().toList()
+                companion.selectAll().toList()
             }
         }
         @Suppress("UNCHECKED_CAST")
@@ -118,7 +145,9 @@ internal class ExposedDao<T : Any>(
     override suspend fun find(username: String): T? {
         val result = withContext(Dispatchers.IO) {
             transaction(this@ExposedDao.database) {
-                AdminUser.find { AdminUsers.username eq username }.firstOrNull()
+                AdminUsers.select(AdminUsers.username, AdminUsers.password).where { AdminUsers.username eq username }
+                    .withDistinct()
+                    .map { it[AdminUsers.username] to it[AdminUsers.password] }.firstOrNull()
             }
         }
         @Suppress("UNCHECKED_CAST")
@@ -134,17 +163,24 @@ internal class ExposedDao<T : Any>(
      * @throws IllegalArgumentException if the entity with the given ID is not found.
      */
     override suspend fun update(data: Map<String, Any>): T {
-        val updatedEntity = withContext(Dispatchers.IO) {
+        @Suppress("UNCHECKED_CAST")
+        val idColumn = companion.primaryKey?.columns?.firstOrNull() as? Column<Any>
+            ?: throw IllegalStateException("No primary key defined for table.")
+        val id = data["id"].toString().toInt()
+
+        @Suppress("UNCHECKED_CAST")
+        val entityId = EntityID(id, idColumn.table as IdTable<Int>)
+        withContext(Dispatchers.IO) {
             transaction(this@ExposedDao.database) {
-                val id = data["id"].toString().toInt()
-                val entityToUpdate = companion.findById(id)
-                    ?: throw IllegalArgumentException("Entity with id $id not found for update.")
-                copyProperties(data, entityToUpdate)
-                entityToUpdate
+                companion.update({ idColumn eq entityId }) {
+                    assignMapToInsertOrUpdate(companion, data, update = it)
+                }
+
             }
         }
+        val entity = this.findById(entityId, false)
         @Suppress("UNCHECKED_CAST")
-        return updatedEntity as T
+        return entity as T
     }
 
     /**
@@ -156,11 +192,14 @@ internal class ExposedDao<T : Any>(
     override suspend fun save(data: Map<String, Any>): T {
         val savedEntity = withContext(Dispatchers.IO) {
             transaction(this@ExposedDao.database) {
-                companion.new { copyProperties(data, this) }
+                companion.insert { assignMapToInsertOrUpdate(companion, data, it) }
             }
         }
+        val idColumn: Column<*>? = companion.columns.find { it.name == "id" } ?: companion.primaryKey?.columns?.first()
+        val id = savedEntity[idColumn!!]
+        val entity = this.findById(id!!, false)
         @Suppress("UNCHECKED_CAST")
-        return savedEntity as T
+        return entity as T
     }
 
     /**
@@ -171,9 +210,9 @@ internal class ExposedDao<T : Any>(
      */
     override suspend fun delete(id: Any): T {
         val obj = withContext(Dispatchers.IO) {
-            transaction(this@ExposedDao.database) { companion.findById(id.toString().toInt()) }
+//            transaction(this@ExposedDao.database) { companion.findById(id.toString().toInt()) }
         }
-        transaction(database) { obj?.delete() } ?: throw IllegalArgumentException("Entity with id $id not found.")
+//        transaction(database) { obj?.delete() } ?: throw IllegalArgumentException("Entity with id $id not found.")
         @Suppress("UNCHECKED_CAST")
         return obj as T
     }
@@ -183,6 +222,6 @@ internal class ExposedDao<T : Any>(
      * This method uses Exposed's SchemaUtils to create the table in the database.
      */
     override suspend fun createTable() {
-        withContext(Dispatchers.IO) { transaction(this@ExposedDao.database) { SchemaUtils.create(companion.table) } }
+        withContext(Dispatchers.IO) { transaction(this@ExposedDao.database) { SchemaUtils.create(companion) } }
     }
 }